dockerized this application, make it work on domains other than localhost
Have spent the past 5 days working to dockerized the  application  while
solving issues  related  to  CORS,  origins  accepted  by  server,  port
mapping, proxies, authentication, subdomains, etc.

The docker setup to run this application in a dockerized environment  is
outside this repo. I managed the setup  of  both  frontend  and  backend
together and decided to keep the docker setup outside of their repos.

Eventually I may decide to (instead of managing the deploy  of  frontend
and backend server together in the same docker-compose) put  the  docker
stuff in each server's repo.