fix bug with noscript in production
The noscript tag does not work with nuxtjs in production mode. This  tag
was used to render some HTML for JS-disabled browsers, but it  does  not
work in production mode.

The solution is to use CSS to conditionally display or hide HTML. If  JS
is loaded, it changes the  CSS  class  of  some  elements,  making  them
invisible. That way, some HTML  will  be  visible  to  only  JS-disabled
users.